agent-1/azure-ai-foundry-agent/requirements.txt
Flask==2.1.1
azure-ai-textanalytics==5.2.0
azure-cosmos==4.2.0
azure-search-documents==11.3.0
requests==2.26.0
python-dotenv==0.19.2
openai==0.27.0
PyGithub==1.55